+++ packages/web/src/pages/_protected/groups/$slug_id/@dashboard/expense-table.tsx
